part 1 move base classes to base.py, start templates module
