+++ composer/models/base.py
@@ -4,17 +4,22 @@
 from __future__ import annotations
 
 import abc
-from typing import Any, Optional, Sequence, Tuple, Union
+import logging
+from typing import TYPE_CHECKING, Any, Mapping, Sequence, Tuple, Union
 
 import torch
 from torch import Tensor
 from torchmetrics import Metric, MetricCollection
-from torchmetrics.classification import Accuracy
 
-from composer.core.types import Batch, BatchPair
-from composer.models.loss import CrossEntropyLoss, soft_cross_entropy
+from composer.core.types import Batch
+from composer.models.nlp_metrics import LanguageCrossEntropyLoss
 
-__all__ = ["ComposerClassifier", "ComposerModel"]
+if TYPE_CHECKING:
+    import transformers
+
+log = logging.getLogger(__name__)
+
+__all__ = ["ComposerModel", "ComposerTransformer"]
 
 
 class ComposerModel(torch.nn.Module, abc.ABC):
@@ -24,7 +29,7 @@ class ComposerModel(torch.nn.Module, abc.ABC):
     implement :meth:`forward` and :meth:`loss`. For full functionality (logging and validation), implement :meth:`metrics`
     and :meth:`validate`.
 
-    See the :doc:`Composer Model walkthrough </composer_model>` for more details.
+    See the :doc:`Composer Model walk through </composer_model>` for more details.
 
     Minimal Example:
 
@@ -198,57 +203,108 @@ class ComposerModel(torch.nn.Module, abc.ABC):
         raise NotImplementedError('Implement validate in your ComposerModel to run validation.')
 
 
-class ComposerClassifier(ComposerModel):
-    """A convenience class that creates a :class:`.ComposerModel` for classification tasks from a vanilla PyTorch model.
-    :class:`.ComposerClassifier` requires batches in the form: (``input``, ``target``) and includes a basic
-    classification training loop with :func:`.soft_cross_entropy` loss and accuracy logging.
+class ComposerTransformer(ComposerModel):
+    """The ComposerModel base interface for Transformers.
+
+    Works with `Hugging Face Transformers <https://huggingface.co/transformers/>`_.
 
     Args:
-        module (torch.nn.Module): A PyTorch neural network module.
-
-    Returns:
-        ComposerClassifier: An instance of :class:`.ComposerClassifier`.
-
-    Example:
-
-    .. testcode::
-
-        import torchvision
-        from composer.models import ComposerClassifier
-
-        pytorch_model = torchvision.models.resnet18(pretrained=False)
-        model = ComposerClassifier(pytorch_model)
+        module (transformers.PreTrainedModel): An instance of PreTrainedModel that
+            contains the forward pass function.
+        config (transformers.PretrainedConfig): The PretrainedConfig object that
+            stores information about the model hyperparameters.
+        tokenizer (transformers.PreTrainedTokenizer): The tokenizer used for this model,
+            necessary to assert required model inputs.
+        gradient_checkpointing (bool, optional): Use gradient checkpointing. Default: ``False``.
     """
 
-    num_classes: Optional[int] = None
-
-    def __init__(self, module: torch.nn.Module) -> None:
+    def __init__(self,
+                 module: transformers.PreTrainedModel,
+                 config: transformers.PretrainedConfig,
+                 tokenizer: transformers.PreTrainedTokenizer,
+                 gradient_checkpointing: bool = False) -> None:
         super().__init__()
-        self.train_acc = Accuracy()
-        self.val_acc = Accuracy()
-        self.val_loss = CrossEntropyLoss()
         self.module = module
+        self.config = config
+        self.tokenizer = tokenizer
+        log.info("Number of parameters in the model: " \
+                 f"{sum(p.numel() for p in module.parameters()):,}")  # type: ignore (thirdparty)
+        log.info("Number of trainable parameters in the model: "
+                 f"{sum(p.numel() for p in module.parameters() if p.requires_grad):,}")  # type: ignore (thirdparty)
 
-        if hasattr(self.module, "num_classes"):
-            self.num_classes = getattr(self.module, "num_classes")
+        # the set of inputs that a model expects
+        # if an algorithm modifies the loss function, it must remove "labels" from this set.
+        self.model_inputs = set(self.tokenizer.model_input_names)
+        self.model_inputs.update({"labels"})
 
-    def loss(self, outputs: Any, batch: BatchPair, *args, **kwargs) -> Tensor:
-        _, targets = batch
-        if not isinstance(outputs, Tensor):  # to pass typechecking
-            raise ValueError("Loss expects input as Tensor")
-        if not isinstance(targets, Tensor):
-            raise ValueError("Loss does not support multiple target Tensors")
-        return soft_cross_entropy(outputs, targets, *args, **kwargs)
+        # define metrics for measurements
+        self.train_loss = LanguageCrossEntropyLoss()
+        self.val_loss = LanguageCrossEntropyLoss()
 
-    def metrics(self, train: bool = False) -> Union[Metric, MetricCollection]:
-        return self.train_acc if train else MetricCollection([self.val_acc, self.val_loss])
+        if gradient_checkpointing:
+            self.module.gradient_checkpointing_enable()  # type: ignore
 
-    def forward(self, batch: BatchPair) -> Tensor:
-        inputs, _ = batch
-        outputs = self.module(inputs)
-        return outputs
+    def loss(self, outputs: Mapping, batch: Batch) -> Union[Tensor, Sequence[Tensor]]:
+        """Computes the loss of the tensor from the output.
 
-    def validate(self, batch: BatchPair) -> Tuple[Any, Any]:
-        _, targets = batch
-        outputs = self.forward(batch)
-        return outputs, targets
+        We don't implement this for the generic Transformer abstraction, since loss
+        functions are model and objective specific. A single model architecture could
+        use a myriad of loss functions which are better left expressed by the user.
+
+        Args:
+            outputs (Mapping): The dictionary output from the model.
+                It could contain the loss as computed by Hugging Face,
+                or algorithms can pop the labels from the input in case
+                they modify the loss function.
+            batch (:class:`~composer.core.types.Batch`): The set of ground truth labels to use to compute the loss against.
+
+        Raises:
+            NotImplementedError: A model-specific and task-specific loss function must be written.
+        """
+        raise NotImplementedError("A model-specific loss function must be written.")
+
+    def forward(self, batch: Batch) -> Mapping:
+        """Runs the forward pass of the model.
+
+        Args:
+            batch (~composer.core.types.Batch): A dictionary of Dict[str, Tensor] of inputs that the
+                model expects, as found in :meth:`.ComposerTransformer.get_model_inputs`.
+
+        Returns:
+            output: A dictionary of model outputs as a ``Mapping``. It will include the loss if `labels` is passed as an input.
+        """
+        if not isinstance(batch, dict):
+            raise ValueError(f'Model expects batch to be a dict, got {type(batch)}')
+
+        for key in self.model_inputs:
+            if key not in batch.keys():
+                raise ValueError(f'Batch missing key: {key}')
+        output = self.module(**batch)  # type: ignore (thirdparty)
+        return output
+
+    def validate(self, batch: Batch) -> Tuple[Mapping, None]:
+        """Runs the validation step.
+
+        Args:
+            batch (~composer.core.types.Batch): a dictionary of Dict[str, Tensor] of inputs
+                that the model expects, as found in :meth:`.ComposerTransformer.get_model_inputs`.
+
+        Returns:
+            Tuple[Mapping, None]: A tuple containing the output from the forward pass.
+                This is fed into directly into the output of :meth:`.ComposerModel.metrics`.
+        """
+        assert self.training is False, "For validation, model must be in eval mode"
+        output = self.forward(batch)
+        return output, None
+
+    def get_model_inputs(self):
+        """Returns a set of inputs that the model expects in the forward pass.
+
+        If an algorithm wants to interact with the model inputs (for instance,
+        popping the labels for a custom loss fn, or adding attention head masks
+        for head pruning, it must access self.set_model_inputs().
+
+        Returns:
+            model_inputs: The set of keys that are expected in the Mapping used to compute the forward pass.
+        """
+        return self.model_inputs
