+++ composer/models/classify_mnist/model.py
@@ -8,8 +8,8 @@ import torch
 import torch.nn as nn
 from torch.nn import functional as F
 
-from composer.models.base import ComposerClassifier
-from composer.models.model_hparams import Initializer
+from composer.models import ComposerClassifier
+from composer.models.base_hparams import Initializer
 
 __all__ = ["Model", "MNIST_Classifier"]
