+++ composer/models/efficientnetb0/efficientnetb0_hparams.py
@@ -6,7 +6,7 @@ from dataclasses import dataclass
 
 import yahp as hp
 
-from composer.models.model_hparams import ModelHparams
+from composer.models.base_hparams import ModelHparams
 
 __all__ = ["EfficientNetB0Hparams"]
