+++ composer/models/efficientnetb0/model.py
@@ -1,7 +1,7 @@
 # Copyright 2021 MosaicML. All Rights Reserved.
 
 """A :class:`.ComposerClassifier` wrapper around the EfficientNet-b0 architecture."""
-from composer.models.base import ComposerClassifier
+from composer.models import ComposerClassifier
 from composer.models.efficientnetb0.efficientnets import EfficientNet
 
 __all__ = ["EfficientNetB0"]
