+++ composer/models/gpt2/model.py
@@ -12,8 +12,8 @@ from typing import TYPE_CHECKING, Mapping, Sequence, Union
 from torch import Tensor
 from torchmetrics import Metric, MetricCollection
 
+from composer.models import ComposerTransformer
 from composer.models.nlp_metrics import Perplexity
-from composer.models.transformer_shared import ComposerTransformer
 
 if TYPE_CHECKING:
     import transformers
