+++ composer/models/resnet/model.py
@@ -6,8 +6,8 @@ from typing import List, Optional
 
 from torchvision.models import resnet
 
-from composer.models.base import ComposerClassifier
-from composer.models.model_hparams import Initializer
+from composer.models import ComposerClassifier
+from composer.models.base_hparams import Initializer
 
 __all__ = ["ComposerResNet"]
