+++ composer/models/resnet/resnet_hparams.py
@@ -6,7 +6,7 @@ from dataclasses import dataclass
 
 import yahp as hp
 
-from composer.models.model_hparams import ModelHparams
+from composer.models.base_hparams import ModelHparams
 from composer.models.resnet.model import ComposerResNet
 
 __all__ = ["ResNetHparams"]
