+++ composer/models/resnet56_cifar10/resnet56_cifar10_hparams.py
@@ -4,7 +4,7 @@
 
 from dataclasses import asdict, dataclass
 
-from composer.models.model_hparams import ModelHparams
+from composer.models.base_hparams import ModelHparams
 
 __all__ = ["CIFARResNetHparams"]
