+++ composer/models/resnet9_cifar10/model.py
@@ -5,7 +5,7 @@
 import torch.nn as nn
 from torchvision.models.resnet import BasicBlock
 
-from composer.models.base import ComposerClassifier
+from composer.models import ComposerClassifier
 
 __all__ = ["ResNet9", "CIFAR10_ResNet9"]
