+++ composer/models/templates/__init__.py
@@ -1,0 +1,3 @@
+from composer.models.templates.classification import ComposerClassifier as ComposerClassifier
+
+__all__ = ["ComposerClassifier"]
