+++ composer/models/timm/model.py
@@ -6,7 +6,7 @@ used to create :class:`.ComposerClassifier`."""
 import textwrap
 from typing import Optional
 
-from composer.models.base import ComposerClassifier
+from composer.models import ComposerClassifier
 
 __all__ = ["Timm"]
