+++ composer/models/vit_small_patch16/model.py
@@ -2,7 +2,7 @@
 
 """Implements ViT-S/16 as a :class:`.ComposerClassifier`."""
 
-from composer.models.base import ComposerClassifier
+from composer.models import ComposerClassifier
 
 __all__ = ["ViTSmallPatch16"]
