+++ tests/algorithms/test_cutmix.py
@@ -8,7 +8,7 @@ import torch.nn.functional as F
 from composer.algorithms import CutMix, CutMixHparams
 from composer.algorithms.cutmix.cutmix import _rand_bbox, cutmix_batch
 from composer.core import Event
-from composer.models.base import ComposerClassifier
+from composer.models import ComposerClassifier
 from tests.common import SimpleConvModel
